+++ libs/core/src/lib/selectors/build-selectors.function.spec.ts
@@ -1,4 +1,5 @@
 import { buildSelectors } from './build-selectors.function';
+import { createSelectorsCache } from './memoize-selectors.function';
 
 const copy = (a: any) => JSON.parse(JSON.stringify(a));
 
@@ -13,6 +14,7 @@ const runTimes = {
   getB1000: 0,
   getA1000MinusB1000: 0,
   getA1000PlusB1000: 0,
+  final: 0,
 };
 
 const selectors = buildSelectors<TestState>()({
@@ -24,9 +26,7 @@ const selectors = buildSelectors<TestState>()({
     runTimes.getB1000++;
     return s.b * 1000;
   },
-})();
-
-const selectors2 = buildSelectors<TestState>()(selectors)({
+})({
   getA1000MinusB1000: s => {
     runTimes.getA1000MinusB1000++;
     return s.getA1000 - s.getB1000;
@@ -35,12 +35,20 @@ const selectors2 = buildSelectors<TestState>()(selectors)({
     runTimes.getA1000PlusB1000++;
     return s.getA1000 + s.getB1000;
   },
+})({
+  final: s => {
+    runTimes.final++;
+    return s.getA1000MinusB1000;
+  },
 })();
 
-describe('combineSelectors', () => {
+const cache1 = createSelectorsCache();
+const cache2 = createSelectorsCache();
+
+describe('buildSelectors', () => {
   const state1 = { a: 4, b: 3, c: 0 };
 
-  const minus = selectors2.getA1000MinusB1000(state1);
+  const minus = selectors.getA1000MinusB1000(state1, cache1);
   const runTimes1m = copy(runTimes);
 
   it('minus should equal 1000', () => {
@@ -52,10 +60,11 @@ describe('combineSelectors', () => {
       getB1000: 1,
       getA1000MinusB1000: 1,
       getA1000PlusB1000: 0, // Not selected yet
+      final: 0,
     });
   });
 
-  const plus = selectors2.getA1000PlusB1000(state1);
+  const plus = selectors.getA1000PlusB1000(state1, cache1);
   const runTimes1p = copy(runTimes);
 
   it('plus should equal 7000', () => {
@@ -67,13 +76,14 @@ describe('combineSelectors', () => {
       getB1000: 1,
       getA1000MinusB1000: 1,
       getA1000PlusB1000: 1,
+      final: 0,
     });
   });
 
   // ================ 1ST STATE CHANGE ====================
   const state2 = { ...state1, c: 1 };
 
-  const minus2 = selectors2.getA1000MinusB1000(state2);
+  const minus2 = selectors.getA1000MinusB1000(state2, cache1);
   const runTimes2m = copy(runTimes);
   it('minus2 should equal 1000', () => {
     expect(minus2).toBe(1000);
@@ -84,10 +94,11 @@ describe('combineSelectors', () => {
       getB1000: 2,
       getA1000MinusB1000: 1, // Same input selector results, so no need to run this again
       getA1000PlusB1000: 1,
+      final: 0,
     });
   });
 
-  const plus2 = selectors2.getA1000PlusB1000(state2);
+  const plus2 = selectors.getA1000PlusB1000(state2, cache1);
   const runTimes2p = copy(runTimes);
   it('plus2 should equal 7000', () => {
     expect(plus2).toBe(7000);
@@ -98,15 +109,16 @@ describe('combineSelectors', () => {
       getB1000: 2,
       getA1000MinusB1000: 1,
       getA1000PlusB1000: 1,
+      final: 0,
     });
   });
 
   // ================ 2ND STATE CHANGE ====================
   const state3 = { ...state2, b: 2 };
 
-  const minus3 = selectors2.getA1000MinusB1000(state3);
+  const minus3 = selectors.getA1000MinusB1000(state3, cache1);
   const runTimes3m = copy(runTimes);
-  it('minus3 should equal 1000', () => {
+  it('minus3 should equal 2000', () => {
     expect(minus3).toBe(2000);
   });
   it('runTimes3m run times', () => {
@@ -115,12 +127,13 @@ describe('combineSelectors', () => {
       getB1000: 3,
       getA1000MinusB1000: 2,
       getA1000PlusB1000: 1, // Has not been selected a 2nd time yet
+      final: 0,
     });
   });
 
-  const plus3 = selectors2.getA1000PlusB1000(state3);
+  const plus3 = selectors.getA1000PlusB1000(state3, cache1);
   const runTimes3p = copy(runTimes);
-  it('plus3 should equal 7000', () => {
+  it('plus3 should equal 6000', () => {
     expect(plus3).toBe(6000);
   });
   it('runTimes3p run times', () => {
@@ -129,6 +142,41 @@ describe('combineSelectors', () => {
       getB1000: 3,
       getA1000MinusB1000: 2,
       getA1000PlusB1000: 2,
+      final: 0,
+    });
+  });
+
+  // =============================== Separate Memoization & Final Selector Memoization =================================
+  // Do something with Store 2. Hopefully it uses a different selector, causing all selectors to run again. Separate memoization
+  selectors.final(state3, cache2);
+  const store2RunTimes1m = copy(runTimes);
+  it('should add full run times for 2nd copy of selectors even though using state3 again', () => {
+    expect(store2RunTimes1m).toEqual({
+      getA1000: 4,
+      getB1000: 4,
+      getA1000MinusB1000: 3,
+      getA1000PlusB1000: 2, // Not selected yet
+      final: 1,
+    });
+  });
+
+  //
+  selectors.final({ a: 10, b: 20, c: 30 }, cache1);
+  selectors.final(state3, cache2);
+  selectors.final({ a: 100, b: 200, c: 300 }, cache1);
+  selectors.final(state3, cache2);
+  selectors.final({ a: 1000, b: 2000, c: 3000 }, cache1);
+  selectors.final({ ...state3, c: 9090 }, cache2); // top input selectors need to run again
+  selectors.final({ a: 10000, b: 2000, c: 30000 }, cache1);
+  selectors.final(state3, cache2); // top input selectors need to run again
+  const runTimes3mAgain = copy(runTimes);
+  it('should run final selector only 4 more times', () => {
+    expect(runTimes3mAgain).toEqual({
+      getA1000: 10,
+      getB1000: 10,
+      getA1000MinusB1000: 7,
+      getA1000PlusB1000: 2, // Still not selected
+      final: 5,
     });
   });
 });